requirements.txt
yfinance
requests
python-dotenv